Progress on: 152180: Graphs don't can't have self-loops https://bugs.eclipse.org/bugs/show_bug.cgi?id=152180
This fixes bugs 114452, 152045, and 152180.

114452:
 Adds ConnectionCounter class to GraphModelFactory to count the number of
connections that go between the same source and destination.
 Adds functionality in createModelFromContentProvider in GraphModelFactory to
adjust the depth on curves when source and dest are repeated.
 Removes the check in GraphModelConnection.createRelationShip(...) for
duplicate connections.

152045:
 GraphModelConnection.createRelationship(...) now allows for one of the source
_or_ dest to be null. If one is null, only a node, with no connection, is
created.

152180:
 In GraphConnectionEditPart, an ArcConnection is created in spite of what
styles are set, if the connection's source and dest are the same.
 Stopped IllegalArgumentException from being thrown in
GraphModelConnection.reconnect(...) if the newSource and newDestination are
equal.
 GraphNodeEditPart now contains two possible anchors: the default, and the
loopAnchor. A new class LoopAnchor has been created in GraphNodeEditPart. It
calculates the reference point based on a point on the outside of the owner's
box, rather than the middle. This prevents loops from being hidden by the
owner.